Remove console size restrictions and catch resize exceptions.
The old warning message logged upon parsing window size command arguments was misleading. "Capping" implies enforcement of a maximum size but the code implemented a minimum size. Regardless of the original code's intent, for better or for worse, the window size is the users' business. Should they wish to have funky window sizes, so be it.

Additionally, threw in try-catch blocks to gracefully handle resize exceptions.