Fix always-blank last row of console output.
The console favored scrolling down instead of moving the prompt from the second-to-last line to the last line, which resulted in a wasted blank line at the bottom of the console.

Additionally, BlankDrawnCommandLine calls were always followed up with a SetCursorPosition to move the cursor to the beginning of the line. Slightly simplified this process by moving the column reset into BlankDrawnCommandLine.