Overhaul ObjectManager implementation.
Both a container that holds all objects and a container that holds zone/object relationships aren't really needed. These have been folded into a single container for only zone/object pairings, which can be queried with LINQ to fulfill the same intent as before. The container has been changed from a Dictionary to a List under the assumption that any one zone will not have duplicate object IDs, as supported by the SQL scripts in the Resources submodule.
Additionally, removed function-local list copies that the functions returned in favor of LINQ statements.

Added a warning for when the filesystem fallback has no directory for the requested zone.

Updated NPC caching to prevent the database from being queried when a zone's NPCs have already been loaded.

Implemented getObjectByID(string, uint) and updated getObjectByID(uint). The function now warns when either the requested object is not found or there are multiple matching objects.